:sparkles: Allow searching plain text and source
